refactor: split auth routes into modular directory structure
Mirrors the bins-routes template (commit 65139f77). The 995-line
routes/auth.ts becomes routes/auth/ with 8 sub-modules:

- index.ts        — composition only
- status.ts       — /status, /demo-login
- invitePreview.ts — /invite-preview
- register.ts     — /register
- session.ts      — /login, /refresh, /logout, /logout-all
- profile.ts      — /me, /profile, /active-location, /password
- account.ts      — DELETE /account, /recover-deletion
- passwordReset.ts — /forgot-password, /reset-password
- oauth.ts        — Google + Apple flows + /oauth/links + DELETE /oauth/link/:provider
- helpers.ts      — shared isLocationMember()

Public surface unchanged. All 24 routes mounted at the same paths,
same request/response contracts, same cookie semantics. Per-route
authenticate middleware preserved (no global router.use(authenticate)
since some routes are public).

Test surface preserved: 1545 server tests still pass. The
appleOAuthNonce stack-walking test imports from auth/oauth.js directly
since the apple-callback layer no longer lives at the top of the
combined router.
